refactor(server): move server subcommand packages under server packages
This syncs directory structure with command structure.
