+++ internal/commands/all/all.go
@@ -8,12 +8,12 @@ import (
 	"github.com/UpCloudLtd/upcloud-cli/internal/commands/ipaddress"
 	"github.com/UpCloudLtd/upcloud-cli/internal/commands/loadbalancer"
 	"github.com/UpCloudLtd/upcloud-cli/internal/commands/network"
-	"github.com/UpCloudLtd/upcloud-cli/internal/commands/networkinterface"
 	"github.com/UpCloudLtd/upcloud-cli/internal/commands/root"
 	"github.com/UpCloudLtd/upcloud-cli/internal/commands/router"
 	"github.com/UpCloudLtd/upcloud-cli/internal/commands/server"
-	"github.com/UpCloudLtd/upcloud-cli/internal/commands/serverfirewall"
-	"github.com/UpCloudLtd/upcloud-cli/internal/commands/serverstorage"
+	serverfirewall "github.com/UpCloudLtd/upcloud-cli/internal/commands/server/firewall"
+	"github.com/UpCloudLtd/upcloud-cli/internal/commands/server/networkinterface"
+	serverstorage "github.com/UpCloudLtd/upcloud-cli/internal/commands/server/storage"
 	"github.com/UpCloudLtd/upcloud-cli/internal/commands/storage"
 	"github.com/UpCloudLtd/upcloud-cli/internal/commands/zone"
 	"github.com/UpCloudLtd/upcloud-cli/internal/config"
