+++ internal/commands/server/firewall/delete.go
