+++ internal/commands/server/firewall/delete_test.go
