+++ internal/commands/server/firewall/show.go
