+++ internal/commands/server/firewall/show_test.go
