+++ internal/commands/server/networkinterface/create.go
