+++ internal/commands/server/networkinterface/create_test.go
