+++ internal/commands/server/networkinterface/delete.go
