+++ internal/commands/server/networkinterface/delete_test.go
