+++ internal/commands/server/networkinterface/modify.go
