+++ internal/commands/server/networkinterface/modify_test.go
