+++ internal/commands/server/networkinterface/network_interface.go
