+++ internal/commands/server/storage/detach.go
